single list of downloaded items
